Add support for specifying included tables to sharding config
The sharding config takes in a list of ignored tables currently. This works well
when the use case calls for including most of the tables. However, if the use
case calls for including a small subset of the tables, then the ignored list
starts becoming very long and unwieldy.

In this commit, we introduce an additional `IncludedTables` config option,
identitcal to `IgnoredTables` in strucuture, but also mutually excusive with it.

The sharding ferry initializes the correct filter based on which one of these
two options are populated, and fails if it finds both to be populated at the
same time.

We also add a couple of tests for verifying this behaviour.

I considered a few ways to do this:
- Passing in a flag to ShardedTableFilter alongside the list the tables
- Passing in the type of the filter to ShardedTableFilter alongside the list of
tables
- Breaking down ShardedTableFilter into two separate types and using the right
one based on the config

The type safety of the solutions increases down the list, but so does the
complexity. I initially started with option 3, thinking it to be the most
robust, but the complexity introduced by the extra layer of abstraction quickly
became hard to justify. As we only have one possible "seam" to break
ShardedTableFilter by, I abandoned this approach till we found a second
seam (YAGNI).

Passing in a boolean flag was the simplest, but also mandated that one of the
two filters become the "default" for naming the flag itself (i.g. `isIgnore` vs
`isInclude`).

The second approach was a good mix of simplicity and robustness as we have named
enum values to work with rather than mysterious booleans, so I went with that in
the end.

Co-authored-by: Shiv Nagarajan <shiv.nagarajan@shopify.com>